# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Atlas Global Resilience Corp. API
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInputLang } from './contactInputLang';

export interface ContactInput {
  name: string;
  organization: string;
  email: string;
  country: string;
  subject: string;
  needType: string;
  message: string;
  lang?: ContactInputLang;
}