# lib/api-zod/src/generated/types/insight.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Atlas Global Resilience Corp. API
 * OpenAPI spec version: 0.1.0
 */

export interface Insight {
  id: number;
  titleFr: string;
  titleEn: string;
  summaryFr: string;
  summaryEn: string;
  bodyFr: string;
  bodyEn: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  publishedAt: string;
  featured: boolean;
  createdAt?: string;
}