# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Atlas Global Resilience Corp. API
 * OpenAPI spec version: 0.1.0
 */
import type { Insight } from './insight';
import type { Project } from './project';

export interface SiteStats {
  totalProjects: number;
  totalServices: number;
  totalInsights: number;
  totalCountries: number;
  featuredProjects: Project[];
  recentInsights: Insight[];
}